# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Latter House Life Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trackers/medications:
    get:
      operationId: listMedications
      tags: [trackers]
      summary: List all medications
      responses:
        "200":
          description: List of medications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medication"
    post:
      operationId: createMedication
      tags: [trackers]
      summary: Add a medication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMedicationBody"
      responses:
        "201":
          description: Created medication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
  /trackers/medications/{id}:
    put:
      operationId: updateMedication
      tags: [trackers]
      summary: Update a medication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMedicationBody"
      responses:
        "200":
          description: Updated medication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
    delete:
      operationId: deleteMedication
      tags: [trackers]
      summary: Delete a medication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /trackers/financial:
    get:
      operationId: listFinancialEntries
      tags: [trackers]
      summary: List financial entries
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: string
          description: Filter by month (YYYY-MM)
      responses:
        "200":
          description: List of financial entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FinancialEntry"
    post:
      operationId: createFinancialEntry
      tags: [trackers]
      summary: Add a financial entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFinancialEntryBody"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialEntry"
  /trackers/financial/{id}:
    delete:
      operationId: deleteFinancialEntry
      tags: [trackers]
      summary: Delete a financial entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /trackers/goals:
    get:
      operationId: listGoals
      tags: [trackers]
      summary: List goals
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [trackers]
      summary: Create a goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoalBody"
      responses:
        "201":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /trackers/goals/{id}:
    put:
      operationId: updateGoal
      tags: [trackers]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGoalBody"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [trackers]
      summary: Delete a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
...
  schemas:
...
    Medication:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        dose:
          type: string
          nullable: true
        times:
          type: string
          nullable: true
        asNeeded:
          type: boolean
        notes:
          type: string
          nullable: true
        refillDate:
          type: string
          nullable: true
        doctor:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - name
        - asNeeded
        - createdAt
    CreateMedicationBody:
      type: object
      properties:
        name:
          type: string
        dose:
          type: string
          nullable: true
        times:
          type: string
          nullable: true
        asNeeded:
          type: boolean
          nullable: true
        notes:
          type: string
          nullable: true
        refillDate:
          type: string
          nullable: true
        doctor:
          type: string
          nullable: true
      required:
        - name
    UpdateMedicationBody:
      type: object
      properties:
        name:
          type: string
        dose:
          type: string
          nullable: true
        times:
          type: string
          nullable: true
        asNeeded:
          type: boolean
          nullable: true
        notes:
          type: string
          nullable: true
        refillDate:
          type: string
          nullable: true
        doctor:
          type: string
          nullable: true
...
    FinancialEntry:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        type:
          type: string
        amount:
          type: string
        description:
          type: string
        category:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - date
        - type
        - amount
        - description
        - createdAt
    CreateFinancialEntryBody:
      type: object
      properties:
        date:
          type: string
        type:
          type: string
        amount:
          type: string
        description:
          type: string
        category:
          type: string
          nullable: true
      required:
        - date
        - type
        - amount
        - description
    Goal:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        targetDate:
          type: string
          nullable: true
        progress:
          type: string
        completed:
          type: boolean
        milestones:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - progress
        - completed
        - createdAt
        - updatedAt
    CreateGoalBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        targetDate:
          type: string
          nullable: true
        milestones:
          type: string
          nullable: true
      required:
        - title
    UpdateGoalBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        targetDate:
          type: string
          nullable: true
        progress:
          type: string
          nullable: true
        completed:
          type: boolean
          nullable: true
        milestones:
          type: string
          nullable: true
...